+++ R/ctrLoadQueryIntoDb.R
@@ -16,9 +16,8 @@
 #' \link{ctrGetQueryUrl} or \link{dbQueryHistory},
 #' or an `_id` in the format of one of the trial registers,
 #' or, together with \code{register}, a string with query
-#' elements of a search URL.
-#' The query details are recorded in the \code{collection} for
-#' later use to update records.
+#' elements of a search URL. The query details are recorded in the
+#' \code{collection} for later use, e.g. to update records.
 #' For "CTIS", \code{queryterm} can be an empty string to obtain
 #' all trial records. For automatically copying the user's
 #' query of a register in a web browser to the clipboard, see
@@ -63,15 +62,24 @@
 #' Used with "CTGOV2", "ISRCTN" and "CTIS" (for "EUCTR", all documents
 #' are downloaded since they are few and have non-canonical filenames.)
 #'
-#' @param euctrresults If \code{TRUE}, also download available
+#' @param euctrresults If \code{TRUE}, also load available
 #' results when retrieving and loading trials from EUCTR. This
 #' slows down this function. (For "CTGOV2" and "CTIS",
 #' available results are always retrieved and loaded into the
 #' collection.)
 #'
-#' @param euctrresultshistory If \code{TRUE}, download results and also
+#' @param euctrresultshistory If \code{TRUE}, load results and also
 #' the available history of results publication in "EUCTR."
 #' This somewhat time-consuming. Default is \code{FALSE}.
+#'
+#' @param euctrprotocolsall If \code{TRUE}, load all available records of
+#' protocol-related data (that is, versions from all EU Member States and any
+#' third country where the trial is conducted); if \code{FALSE}, only a single
+#' record per trial is loaded, to accelerate loading. Default is \code{TRUE},
+#' but only for backwards consistency; for new collections, \code{FALSE} is
+#' the recommended setting, unless there are questions about differences
+#' between Member States' protocol versions of a trial such as dates or outcomes
+#' of an authorisation decision or an ethics opinion, global status and end.
 #'
 #' @param ctgov2history For trials from CTGOV2, retrieve historic
 #' versions of the record. Default is \code{FALSE}, because this
@@ -114,9 +122,7 @@
 #' @param verbose If \code{TRUE}, prints additional information
 #' (default \code{FALSE}).
 #'
-#' @param ... Do not use (capture deprecated parameters).
-#'
-#' @return A list with elements
+#' @returns A list with elements
 #' `n` (number of trial records newly imported or updated),
 #' `success` (a vector of _id's of successfully loaded records),
 #' `failed` (a vector of identifiers of records that failed to load)
@@ -180,6 +186,7 @@ ctrLoadQueryIntoDb <- function(
     forcetoupdate = FALSE,
     euctrresults = FALSE,
     euctrresultshistory = FALSE,
+    euctrprotocolsall = TRUE,
     ctgov2history = FALSE,
     ctishistory = FALSE,
     documents.path = NULL,
@@ -188,29 +195,9 @@ ctrLoadQueryIntoDb <- function(
     annotation.mode = "append",
     only.count = FALSE,
     con = NULL,
-    verbose = FALSE, ...) {
+    verbose = FALSE) {
 
   ## check params
-
-  # - deprecated
-  params <- list(...)
-  if (!is.null(params$euctrresultspdfpath)) {
-    warning("Parameter 'euctrresultspdfpath' is deprecated, ",
-            "use 'documents.path'",
-            call. = FALSE
-    )
-    documents.path <- params$euctrresultspdfpath
-  }
-  if (!is.null(params$euctrresultsfilespath)) {
-    warning("Parameter 'euctrresultsfilespath' is deprecated, ",
-            "use 'documents.path'",
-            call. = FALSE
-    )
-    documents.path <- params$euctrresultsfilespath
-  }
-  if (!is.null(params$parallelretrievals)) {
-    warning("Parameter 'parallelretrievals' is deprecated and ignored")
-  }
 
   # - parameters consistent
   if (!is.null(querytoupdate) && !is.null(queryterm)) {
@@ -347,6 +334,7 @@ ctrLoadQueryIntoDb <- function(
     register = register,
     euctrresults = euctrresults,
     euctrresultshistory = euctrresultshistory,
+    euctrprotocolsall = euctrprotocolsall,
     ctgov2history = ctgov2history,
     ctishistory = ctishistory,
     documents.path = documents.path,