+++ R/f_startDate.R
@@ -12,7 +12,7 @@
 #' prints fields needed in `df` for calculating this trial concept, which can
 #' be used with \link{dbGetFieldsIntoDf}.
 #'
-#' @return data frame with columns `_id` and `.startDate`, a date.
+#' @returns data frame with columns `_id` and `.startDate`, a date.
 #'
 #' @export
 #'
@@ -74,7 +74,7 @@ f.startDate <- function(df = NULL) {
   #### calculate ####
 
   # check generic, do not edit
-  fctChkFlds(names(df), fldsNeeded)
+  df <- fctChkFlds(df, fldsNeeded)
 
   # helper function
   `%>%` <- dplyr::`%>%`