+++ R/f_trialObjectives.R
@@ -14,7 +14,7 @@
 #' prints fields needed in `df` for calculating this trial concept, which can
 #' be used with \link{dbGetFieldsIntoDf}.
 #'
-#' @return data frame with columns `_id` and `.trialObjectives`, which is
+#' @returns data frame with columns `_id` and `.trialObjectives`, which is
 #' a string with letters separated by a space, such as
 #' E (efficacy, including cure, survival, effectiveness);
 #' A (activity, including reponse, remission, seroconversion);
@@ -111,7 +111,7 @@ f.trialObjectives <- function(df = NULL) {
   #### calculate ####
 
   # check generic, do not edit
-  fctChkFlds(names(df), fldsNeeded)
+  df <- fctChkFlds(df, fldsNeeded)
 
   # apply nested function which provides values for each register
   # therefore the following code needs to check against register