+++ inst/tinytest/ctrdata_isrctn.R
@@ -231,24 +231,55 @@ expect_true(
   res[, "annotation", drop = TRUE] == "just_this" &&
     res[, "_id", drop = TRUE] == "98918118")
 
-# clean up
-rm(res)
-
 # test
 suppressWarnings(
   suppressMessages(
-    tmpDf <- dbGetFieldsIntoDf(
+    res <- dbGetFieldsIntoDf(
       fields = c(
         "participants.totalFinalEnrolment"
       ), con = dbc)))
 #
 expect_equivalent(
-  sapply(tmpDf, typeof),
+  sapply(res, typeof),
   c("character", "integer")
 )
 
+expect_message(
+  res <- suppressWarnings(
+    dbGetFieldsIntoDf(
+      fields = c("_id"),
+      con = dbc)),
+  "calling dbFindIdsUniqueTrials")
+
+# test
+expect_true(length(res) >= 14L)
+
+expect_warning(
+  res <- dbGetFieldsIntoDf(
+    fields = c("_id"),
+    dbc),
+  "use the named parameter con")
+
+# test
+expect_true(length(res) >= 14L)
+
+if (any(class(dbc) %in% c("src_sqlite", "src_postgres"))) {
+
+  expect_message(
+    res <- suppressWarnings(dbGetFieldsIntoDf(
+      fields = c("ctrname", paste0("f_", 1:1000))[-1],
+      calculate = "f.statusRecruitment",
+      con = dbc)),
+    "switching to iterating")
+
+  # test
+  expect_true(ncol(res) == 2L)
+  expect_true(nrow(res) >= 14L)
+
+}
+
 # clean up
-rm(tmpDf)
+rm(res)
 
 #### dbFindFields ####
 
@@ -318,18 +349,16 @@ rm(res)
 
 #### ctrLoadQueryIntoDb documents ####
 
-if (interactive()) {
+tmpDir <- newTempDir()
+unlink(tmpDir, recursive = TRUE)
+on.exit(unlink(tmpDir, recursive = TRUE), add = TRUE)
 
-  tmpDir <- newTempDir()
-  on.exit(unlink(tmpDir, recursive = TRUE), add = TRUE)
-
-  expect_message(
-    ctrLoadQueryIntoDb(
-      queryterm = "https://www.isrctn.com/search?q=alzheimer",
-      con = dbc,
-      documents.path = tmpDir,
-      documents.regexp = ".*"
-    ),
-    "Newly saved [0-9]+ document[(]s[)] for [0-9]+ trial"
-  )
-}
+expect_message(
+  ctrLoadQueryIntoDb(
+    queryterm = "https://www.isrctn.com/search?q=alzheimer",
+    con = dbc,
+    documents.path = tmpDir,
+    documents.regexp = ".*"
+  ),
+  "Newly saved [0-9]+ document[(]s[)] for [0-9]+ trial"
+)