+++ inst/tinytest/test_ctrdata_function_various.R
@@ -209,6 +209,18 @@ tmpTest <- suppressMessages(
 # test
 expect_true("data.frame" %in% class(tmpTest))
 
+tmpTest <- suppressMessages(
+  ctrGetQueryUrl(
+    url = "https://classic.clinicaltrials.gov/ct2/results?type=Intr&cntry=CL&city=Santiago"))
+# via web ui:
+# https://clinicaltrials.gov/search?country=Chile&aggFilters=studyType:int&locStr=Santiago,%20Chile&state=Santiago%20Metropolitan%20Region
+
+# test
+expect_equal(
+  tmpTest$`query-term`,
+  "country=Chile&city=Santiago&aggFilters=studyType:int"
+)
+
 # test
 expect_warning(
   ctrGetQueryUrl(
@@ -420,36 +432,6 @@ expect_equal(
 #### ctrLoadQueryIntoDb parameters ####
 
 if (checkInternet()) {
-
-  # test
-  expect_warning(
-    ctrLoadQueryIntoDb(
-      queryterm = "somethingnonexisting",
-      register = "EUCTR",
-      euctrresultsfilespath = "something",
-      only.count = TRUE
-    ), "deprecated"
-  )
-
-  # test
-  expect_warning(
-    ctrLoadQueryIntoDb(
-      queryterm = "somethingnonexisting",
-      register = "EUCTR",
-      euctrresultspdfpath = "something",
-      only.count = TRUE
-    ), "deprecated"
-  )
-
-  # test
-  expect_warning(
-    ctrLoadQueryIntoDb(
-      queryterm = "somethingnonexisting",
-      register = "EUCTR",
-      parallelretrievals = 99L,
-      only.count = TRUE
-    ), "ignored"
-  )
 
 }